Update FakeXRDeviceInterface to match webxr-test-api
This updates the FakeXRDeviceInterface to support newly added methods in
the webxr-test-api:
* simulateVisibilityChange
* setBoundsGeometry
* setLocalToFloorLevelTransform
* clearLocalToFloorLevelTransform

and supports the setting of boundsCoordinates and
localToFloorLevelTransform in FakeXRDeviceInit.

The Bounds and LocalToFloorLevelTransform functions already existed on
internal tests, and as such a few of the tests that were ready were
ported to external wpts.

Note that simulateInputSourceConnection and disconnect will be
implemented by future work.

Bug:979316
Change-Id: Idfb8bb631bfc1f1cb13b571a90b1a41b4a6fe518