CSS @media (width: *px) may not work if devicePixelRatio is greater than 1.
When devicePixelRatio is greater than 1, the viewport size of the CSS media
query may appear as non-integers. This can happen in Chrome for Android if
the UseZoomForDSF flag is enabled.

For example:
<meta name="viewport" content="width=device-width, initial-scale=1">
<style>
    @media (width: 393px) {
        p {
            color: red;
        }
    }
</style>
<p>text</p>

If the device width is 1081 physical pixels and devicePixelRatio is 2.75. The
viewport width of the CSS media query will be 393.091px.

Change-Id: I25ee4da1ebdbd57202fa8c5e360138a8058f578f